update readme with fee instructions
fix bug: register the "empty" account

actually, it's interesting. even with no "beneficiary" argument supplied, the [0x00, 0x00] account will be credited, and this can (probably) be spent from using a secret key of 0. so in this way miners could make "raking" transactions. probably'd be better to have a special "rake" method in the contract, so that the extra gas doesn't have to be spent.

or maybe it's better this way, so that the miner's identity can be kept private... intriguing.

final docs update

mirrored at https://eprint.iacr.org/2020/293

Update README.md

Updated README.md with examples and more code to improve intuitiveness